pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "huggingface-hub>=0.35.3",
    "python-dotenv>=1.1.1",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]
